Added Gizmo Icons
- Fixed alot of warning in Yaml-cpp
- Fixed bugs in Log.cpp
- Added GizmoIcons to EditorResources
- Changed XYZ font in TranslationComponentUI
- Added the option to change the font when pushing temporary font with ImGuiFontsLibrary
- Fixed bug in imgui_demo.cpp
- Added RAII wrappers for ImGui::Push/Pop style/color vars
- Fixed bug in Renderer::CreateEnvironmentMap
- Added Gizmo toolbar to viewport

+++ Aurora/dependencies/yaml-cpp/src/emitterutils.cpp
@@ -222,8 +222,8 @@ std::pair<uint16_t, uint16_t> EncodeUTF16SurrogatePair(int codePoint) {
   const uint32_t leadOffset = 0xD800 - (0x10000 >> 10);
 
   return {
-    leadOffset | (codePoint >> 10),
-    0xDC00 | (codePoint & 0x3FF),
+    (uint16_t)(leadOffset | (codePoint >> 10)),
+    (uint16_t)(0xDC00 | (codePoint & 0x3FF)),
   };
 }
